Return errors from run function
`Result<(), Box<dyn Error>>`

`()` => unit type, using `()` is idiomatic Rust to indicate that we're
calling this function only for its side effects, it does not return a
value we need.
`Box<dyn Error>` => trait object, this function will return a type that
implements the Error trait, without having to specify what particular
type that is. This gives us flexibility to return error values that may
be of different types in different error cases, dyn means "dynamic".