Add failing test case for search
We need an explicit lifetime (`'a`) defined in the signature of search
for the contents argument, and with the return value.

The lifetime parameters specify which argument lifetime is connected to
the lifetime of the return value. Here we indicate that the returned
vector should contain string slices that reference slices of the
argument `contents`, and not the argument `query`.

So we tell Rust that the data returned by the search function will live
as long as the data passed into the search function in the contents
argument.
The data references _by_ a slice needs to be valid for the reference to
be valid; if the compiler assumes we're making string slices of query
instead of contents, it will do its safety checking incorrectly.